ref: moved sessions to its own package
